쇼핑몰 개발노트(09/00) - 새탭에서 이미지 열기시 브라우저로 전달되는 데이터가 text/html 형태로 전송되어 새로 열린 탭에서는 깨진 글자로만 보이는 현상이 있어 컨트롤러에서 produces 를 image/jpeg 로 세팅하여 이미지가 잘 보이도록 수정 - AWS EC2 배포가능 - 회원정보 수정시(닉네임) 본인이 작성한 게시글 수정 및 삭제 불가능한 현상 수정 - 필요없는 컬럼, 클래스 내 필드 삭제 - HTML에 필요없이 존재하던 단어 삭제
 --     Prev dev Note        --
 CURD 게시판 개발노트(08/06)
 - 첨부파일 업로드 & 다운로드 기능 로직 수정
  - 하나의 파일만 업로드될 경우 업로드된 파일이 보이지 않았음.
  - 하나의 파일만 업로드해도 다운로드 가능하도록, 첨부파일 없어도 게시글 정상등록 되도록 수정
 ------------------------------------------------------------
 - 검색결과 페이징 및 페이징 버튼 개발 모두 완료
 - summernote 이미지 ajax 기능 개발 완료
 - 첨부파일 업로드 & 다운로드 기능 개발 완료
 - 댓글/대댓글 기능 개발예정 + viewForm 디자인 수정예정
 ------------------------------------------------------------
 CRUD 게시판 개발노트( ~ 08/11)
 - 회원가입 / 로그인 기능 개발 완료
 - 회원가입 / 로그인 Validation 완료
 - 게시판 및 게시글 add, edit 시 Validation 완료
 - 본인글만 삭제수정 가능, 어드민 모든게시글 권한있음 인터셉터 통한 권한코딩 완료
 - 게시판 접속 시 인터셉터 통한 LoginCheck 기능 개발 완료
 - 잘못된 페이지 또는 악의적인 접근 차단 위해 에러페이지 제공, 예외처리
 ------------------------------------------------------------
 쇼핑몰 개발노트( ~ 08/16)
 - 어드민 페이지 개발 완료
 - 어드민 페이지 통해 회원 권한 조절기능 개발 완료
 - 어드민 페이지 상품 CRUD 기능 개발 완료
 - 어드민 페이지 상품 CRUD Validation 완료
 - Index 페이지 및 LoginIndex 페이지에 상품 노출 연동 완료
 - 상품상세 페이지 개발 완료
 - 상품상세의 관련된 상품(카테고리 기준) 노출 기능 개발 완료
 - 카테고리기준으로 상품별 노출기능 개발완료(Index 의 카테고리탭)
 ------------------------------------------------------------
 쇼핑몰 개발노트(08/17)
 - 로그인/비로그인 화면 분리된 페이지 추가
 - discountProd 페이지에 가격이 반대로 노출되는 현상 수정
 - loginDiscountProd 페이지에 가격이 반대로 노출되는 현상 수정
 ------------------------------------------------------------
 쇼핑몰 개발노트(08/18)

 - 장바구니 페이지 개발 완료
 - 장바구니 기능 개발 완료
 - 장바구니가 보이는 모든 페이지에 장바구니 기능 적용 완료
 ------------------------------------------------------------
 쇼핑몰 개발노트(08/20)

 - 장바구니 상품 없을때 상품없음 문구 노출되도록 수정
 - 장바구니에 담은 상품이 재고가 없다면 재고없음 문구노출되도록 수정
 - 주문페이지 작성완료
 - 주문페이지 다음주소API 적용 완료 및 주문결제기능 개발 완료
 - 포인트결제가 가능하도록 멤버데이터베이스에 포인트 컬럼 추가(회원가입시 디폴트값은 0)
 - 주문 -> 주문금액 포인트 차감 -> 장바구니 목록 삭제 프로세스 적용 완료
 - 주문 -> 주문금액 포인트 차감 -> 상품재고감소 -> 장바구니 목록삭제 프로세스 적용완료
  - 장바구니가 보이는 모든 페이지에 장바구니 기능 적용 완료
  ------------------------------------------------------------
 쇼핑몰 개발노트(08/22)
 - 주문화면 validation 완료
 - 유저 마이페이지 화면 개발완료
 - 유저 개인정보 확인 및 개인정보 진입 프로세스(비밀번호 확인 후 입장) 개발완료
 - 유저 개인정보 수정 기능 개발완료
 - 유저 개인정보 수정 필드 validation 완료
 - 유저 주문내역보기 기능 개발완료
 -------------------------------------------------------------
 쇼핑몰 개발노트(08/25)
 - 마이페이지 validation 완료
 - 쇼핑몰 상단 nav 디자인 개선 & 잔여 포인트 노출영역 추가
 - 결제모듈 연동 완료
  -------------------------------------------------------------
 쇼핑몰 개발노트(08/29)
 - 최종 푸시
 - 결제모듈 연동 및 테스트 완료
 - 페이지 footer 글귀 수정 / 카테고리 수정 등 HTML 다듬기 완료
 - AWS EC2 RDS 연동 위해 properties 내용 수정
   -------------------------------------------------------------
 쇼핑몰 개발노트(08/31)
 - AWS EC2 계정문제로 EC2 배포 불발로 인해 다른 호스팅 서비스 모색중
 - 상품주문 후 mybatis binding 에러로 주문내역페이지로 리다이렉트 하지 못하는 현상 수정
 - 상품 썸네일 선택시 상품상세로 이동할 수 있도록 수정
 - 장바구니에서 상품명 선택시 상품상세로 이동할 수 있도록 수정

+++ src/main/java/forum/board/controller/BoardController.java
@@ -83,9 +83,7 @@ public class BoardController {
        if(bindingResult.hasErrors())
        {
            model.addAttribute("member",loginMember);
-           for (FieldError fieldError : bindingResult.getFieldErrors()) {
-               System.out.println("fieldError = " + fieldError);
-           }
+           model.addAttribute("categoryType",CategoryType.CATEGORY_TYPE);
            return "addForm";
        }
 
@@ -136,6 +134,8 @@ public class BoardController {
     {
         Item item = itemService.findById(itemId);
 
+        // 로그인한 멤버의 권한이 USER 이고 if 조건 분기에 따라 본인 게시글이면 viewForm 으로 이동하고 아니면 viewFomOther 로 이동하도록 설계했다.
+        // 로그인한 멤버의 권한이 USER 가 아니라면 ADMIN 이므로, ADMIN 의 권한에 따라 모든 게시글의 CURD 가 가능하므로 바로 viewForm 으로 이동하도록 설계했다.
 
         if(loginMember.getRole().equals("USER")) {
             if (!(loginMember.getMemberId().equals(item.getWriterId()))) {
@@ -159,12 +159,17 @@ public class BoardController {
     public String editForm(@PathVariable Long itemId, @SessionAttribute(name = SessionConst.LOGIN_MEMBER,required = false) LoginMember loginMember, Model model)
     {
         Item item = itemService.findById(itemId);
-        model.addAttribute("form",new ItemUpdateForm(item.getItemId(),item.getTitle(),item.getContent()));
-        model.addAttribute("member",loginMember);
-        model.addAttribute("categoryType",CategoryType.CATEGORY_TYPE);
 
+        // 본인의 게시글이 아닌 다른사람이 작성한 게시글을 수정하려고 url 을 변경하면 수정이 가능했기때문에, 이를 방지하고자 if 조건을 걸어서 본인의 게시글만 수정페이지로 넘어올 수 있도록 수정하였다.
+        // 어드민은 모든 게시글에 대한 CRUD 권한을 가지므로, or 조건으로 권한체크 조건을 같이 걸어두었다.
+        if(loginMember.getMemberId() == item.getWriterId() || loginMember.getRole().equals("ADMIN")) {
+            model.addAttribute("form", new ItemUpdateForm(item.getItemId(), item.getTitle(), item.getContent()));
+            model.addAttribute("member", loginMember);
+            model.addAttribute("categoryType", CategoryType.CATEGORY_TYPE);
+            return "editForm"; //editForm 에서 "수정완료" 버튼을 누르면 아래의 컨트롤러의 url 로 이동한다.
+        }else
+            return "redirect:/forum/freeBoard/viewForm/{itemId}";
 
-        return "editForm"; //editForm 에서 "수정완료" 버튼을 누르면 아래의 url 로 이동한다.
     }
 
     @PostMapping("/forum/freeBoard/editForm/{itemId}") //게시글이 수정되면, 수정된 게시글 상세보기 페이지로 이동
@@ -189,11 +194,16 @@ public class BoardController {
     }
 
     @GetMapping("/forum/freeBoard/viewForm/delete/{itemId}") // 게시글 상세보기 페이지의 "게시글 삭제" 버튼을 누르면 이 링크로 요청이 온다.
-    public String removeItem(@PathVariable Long itemId)
+    public String removeItem(@PathVariable Long itemId, @SessionAttribute(name = SessionConst.LOGIN_MEMBER,required = false)LoginMember loginMember)
     {
-        itemService.deleteItem(itemId);
-
-        return "redirect:/forum/freeBoard";
+        // 본인 이외 남이 작성한 게시글을 삭제할 수 있는 현상이 있는데 이는 세션에 있는 회원의 정보와 게시글정보의 일치점을 찾는 체크로직이 없어서이다. 체크 로직을 넣겠다.
+        // 본인이 작성한 게시글만 삭제가 가능하며, 어드민은 모든 게시글에대해 CRUD 권한을 가지므로 or 조건으로 권한체크 조건을 같이 걸어두었다.
+        Item item = itemService.findById(itemId);
+        if(loginMember.getMemberId() == item.getWriterId() || loginMember.getRole().equals("ADMIN")) {
+            itemService.deleteItem(itemId);
+            return "redirect:/forum/freeBoard";
+        }else
+            return "redirect:/forum/freeBoard/viewForm/{itemId}";
     }
 
     @GetMapping("/forum/freeBoard/search")
